// src/data/api/generated/models/listRole.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Lupira Tasks API
 * Task and command processing backend for Lupira. Authenticate with a Bearer token issued by the OIDC provider (Authentik).
 * OpenAPI spec version: v1
 */

export type ListRole = typeof ListRole[keyof typeof ListRole];


export const ListRole = {
  Owner: 'Owner',
  Editor: 'Editor',
  Viewer: 'Viewer',
} as const;